Solve 9 of 13
Passed:You can send a DELETE request to /api/replies/{board}
and pass along the thread_id, reply_id, & delete_password.
Returned will be the string incorrect password or success.
On success, the text of the reply_id will be changed to [deleted].

Debugging is a bitch sometimes. I just forgot to return the result.