Solve 11 and 12 of 13 Challenges
You can send a PUT request to /api/threads/{board}
and pass along the thread_id. Returned will be
the string reported. The reported value of the thread_id
will be changed to true

You can send a PUT request to /api/replies/{board} and pass along
the thread_id & reply_id. Returned will be the string reported.
The reported value of the reply_id will be changed to true.